requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==0.7.16
anyio==4.6.2
archspec==0.2.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
boltons==24.0.0
Brotli==1.1.0
brotlicffi==1.1.0.0
certifi==2024.8.30
cffi==1.17.1
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
conda==24.7.1
conda-libmamba-solver==24.11.1
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
docutils==0.21.2
fenics-dijitso==2019.1.0
fenics-dolfin==2019.1.0
fenics-ffc==2019.1.0
fenics-fiat==2019.1.0
fenics-ufl==2019.1.0
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.6.1
gmsh==4.13.1.dev1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
libmambapy==2.0.5
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
meshio==5.3.5
mpi4py==4.0.1
mpmath==1.3.0
multidict==6.1.0
netCDF4==1.7.1.post2
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pkgconfig==1.5.5
platformdirs==4.3.6
pluggy==1.5.0
psutil==6.0.0
pybind11==2.13.6
pycares==4.4.0
pycosat==0.6.3
pycparser==2.22
pyg_lib==0.4.0+pt26cu124
Pygments==2.18.0
pynvml==11.5.3
pyOpenSSL==24.2.1
pyparsing==3.1.4
PyStemmer==2.2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.8.1
ruamel.base==1.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.7
scikit-build==0.18.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.1.post0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-autodoc-typehints==2.4.4
sphinx_rtd_theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.8
sphinxcontrib-serializinghtml==1.1.10
sphinxcontrib-websupport==1.2.7
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.6.0+cu124
torch-geometric==2.6.1
torch_cluster==1.6.3+pt26cu124
torch_scatter==2.1.2+pt26cu124
torch_sparse==0.6.18+pt26cu124
torchaudio==2.6.0+cu124
torchdiffeq==0.2.5
torchvision==0.21.0+cu124
tqdm==4.66.5
triton==3.2.0
trustme==1.1.0
truststore==0.9.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
wheel==0.45.1
yarl==1.13.1
zstandard==0.23.0
